requirements.txt
absl-py==0.4.1
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.4.1
bleach==2.1.4
blinker==1.3
boto==2.38.0
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
cloud-init==18.3
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
cycler==0.10.0
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.2.3
Flask==1.0.2
Flask-Cors==3.0.6
gast==0.2.0
google-api-core==1.4.0
google-auth==1.5.1
google-cloud==0.34.0
google-cloud-bigquery==1.5.0
google-cloud-core==0.28.1
google-cloud-storage==1.12.0
google-compute-engine==2.8.2
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.14.2
h5py==2.8.0
html5lib==1.0.1
idna==2.7
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonpatch==1.10
jsonpointer==1.9
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
language-selector==0.1
lxml==3.5.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mecab-python3==0.7
mistune==0.8.3
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
notebook==5.6.0
numpy==1.15.1
oauthlib==1.0.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pipreqs==0.4.9
prettytable==0.7.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.2.1
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.7.3
python-debian==0.1.27
python-systemd==231
pytz==2018.5
pyxdg==0.25
PyYAML==3.11
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
rsa==4.0
scikit-learn==0.19.2
scipy==1.1.0
screen-resolution-extra==0.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
ssh-import-id==5.5
system-service==0.3
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1.1
traitlets==4.3.2
ufw==0.35
unattended-upgrades==0.1
urllib3==1.23
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
xkit==0.0.0
yarg==0.1.9
